case comment changes, added fluentUI dialog

+++ Hippra/Models/DTO/AddNotificationDto.cs
@@ -11,6 +11,6 @@ namespace Hippra.Models.DTO
         public string ReceiverUserID { get; set; }
         public NotificationType Type { get; set; } = NotificationType.AddedComment;
         public int PostID { get; set; }
-        public int CommentId { get; set; }
+        public long CommentId { get; set; }
     }
 }
